change integer to numeric

+++ backend/server.js
@@ -71,7 +71,7 @@ async function initDb() {
       warehouse_id INTEGER NOT NULL,
       name TEXT NOT NULL,
       category TEXT NOT NULL DEFAULT 'Pieces',
-      quantity INTEGER NOT NULL DEFAULT 0,
+      quantity NUMERIC NOT NULL DEFAULT 0,
       created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
     )
   `);
@@ -81,8 +81,8 @@ async function initDb() {
       id SERIAL PRIMARY KEY,
       product_id INTEGER NOT NULL,
       type TEXT NOT NULL,
-      amount INTEGER NOT NULL,
-      quantity_after INTEGER NOT NULL,
+      amount NUMERIC NOT NULL,
+      quantity_after NUMERIC NOT NULL,
       note TEXT,
       created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
     )
